requirements.txt
# Core Data Processing
apache-airflow==2.7.2
pyspark==3.5.0
kafka-python==2.0.2

# Databases
pymongo==4.6.1
mysql-connector-python==8.2.0
elasticsearch==8.11.1

# AWS
boto3==1.34.34
s3fs==2024.2.0

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.3

# Data APIs
yfinance==0.2.35
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.3

# Email
email-validator==2.1.0

# Monitoring & Logging
python-json-logger==2.0.7

# Data Processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
schedule==1.2.1
pytz==2024.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
